refactor: update dashboard UI and project card components
- Modified dashboard page layout and functionality
- Updated project card component styling and behavior
- Added new task completion interface
