+++ src/app/(routes)/notifications/page.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { getUser } from "@/app/api/api";
+import { getUser } from "@/app/api/userRequests";
 import LoadingPage from "@/app/components/loader";
 import MainLayout from "@/app/components/MainLayout";
 import NotificationCard from "@/app/components/notificationCard";
@@ -7,8 +7,6 @@ import { INotificationProps } from "@/app/interface/INotification";
 import { IUser } from "@/app/interface/IUser";
 import { getAccessTokenFromCookies, parseJwt } from "@/app/utils/utils";
 import { useEffect, useState } from "react";
-
-
 
 export default function NotificationsPage() {
   const [user, setUser] = useState<IUser | null>(null);
