+++ src/app/interface/IProject.ts
@@ -6,4 +6,5 @@ export interface IProject {
     project_id?: string;
     color?: string;
     task_count?: number;
+    completed_tasks?: number;
   }
