+++ src/app/interface/ITaskCompletionTrendData.ts
@@ -1,0 +1,4 @@
+export interface ITaskCompletionTrendData {
+    day: string;
+    completed: number;
+}
